Separated out several files and functions
